Intial work on doorways
This base model has all the faces still intact, and is
a single model. The cutout has been done in a way which
preserves the UVs so therefore there shouldn't be any
textureing issue when using this next to another model.

Note the dimensions of the doorway were carefully measured
as part of the cutting process to be standard measurements
of 2050mm tall by 820mm wide.